NO-JIRA refactor test config
