+++ src/test/session-details-provider/index.js
@@ -1,0 +1,5 @@
+const { SESSION_CONFIRMATION_CODE_URL } = require('./constants')
+
+module.exports = {
+  SESSION_CONFIRMATION_CODE_URL
+}
