+++ src/test/session-details-provider/session-details-app.js
@@ -7,7 +7,7 @@ const config = require('../../config')
 const app = express()
 const { COOKIE_EXPIRES_MILLISECONDS } = require('../../web/server/session/cookie-settings')
 const { CONFIRMATION_CODE_SESSION_PROPERTY } = require('../../web/routes/application/common/constants')
-const { SESSION_DETAILS_PORT, SESSION_DETAILS_PATH } = require('../common/config')
+const { SESSION_DETAILS_PORT, SESSION_DETAILS_PATH } = require('./constants')
 
 const client = redis.createClient(config.redis)
 const store = new RedisStore({ client })
